Dodge Shooter 0.1.2
게임매니저가 이제 몬스터를 랜덤으로 생성함
플레이어가 무적상태인지 더 알기 쉽게 보여짐

+++ Dodge/Assets/Scripts/Bullet.cs
@@ -6,7 +6,7 @@ public class Bullet : MonoBehaviour
 {
 	Rigidbody bulletRigidbody;
 	public float speed;
-	public int damage = 30;
+	public float damage = 30;
 
 	void Start()
 	{
